requirements.txt
streamlit>=1.48.0
pandas>=2.0.0
matplotlib>=3.7.0
tinkoff-investments>=0.2.0b115
ta>=0.10.0
numpy>=1.21.0
plotly>=5.0.0
Pillow>=9.0.0
